meshing/requirements.txt
about-time==4.2.1
alive-progress==3.2.0
ansys-api-fluent==0.3.28
ansys-api-platform-instancemanagement==1.1.0
ansys-api-tools-filetransfer==0.1.1
ansys-fluent-core==0.26.1
ansys-fluent-visualization==0.14.0
ansys-platform-instancemanagement==1.1.2
ansys-tools-filetransfer==0.1.0
ansys-units==0.3.2
beartype==0.19.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
docker==7.1.0
fonttools==4.55.0
grapheme==0.6.0
grpcio==1.68.0
grpcio-health-checking==1.62.3
idna==3.10
importlib_metadata==8.5.0
joblib==1.4.2
kiwisolver==1.4.7
lxml==5.3.0
marshmallow==3.23.1
matplotlib==3.9.2
nltk==3.9.1
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.5
psutil==6.1.0
pyparsing==3.2.0
PySide6==6.7.3
PySide6_Addons==6.7.3
PySide6_Essentials==6.7.3
python-dateutil==2.9.0.post0
pytz==2024.2
pyvista==0.44.1
pyvistaqt==0.11.1
PyYAML==6.0.2
QtPy==2.4.2
regex==2024.11.6
requests==2.32.3
scooby==0.10.0
shiboken6==6.7.3
six==1.16.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vtk==9.3.1
zipp==3.21.0
